+++ controllers/exploreCategory.js
@@ -1,11 +1,11 @@
 const ExploreCategory = require("../models/exploreCategory");
 const shortid = require("shortid");
 const slugify = require("slugify");
-
+const path = require("path");
+const fs = require("fs");
 exports.createExploreCategory = (req, res) => {
   try {
     const { imageTitle, imageAltText, categoryId } = req.body;
-
     const exploreCat = new ExploreCategory({
       imageTitle,
       imageAltText,
@@ -13,7 +13,7 @@ exports.createExploreCategory = (req, res) => {
       createdBy: req.user._id,
     });
     if (req.file) {
-      exploreCat.image = req.file.location;
+      exploreCat.image = process.env.API + "/public/" + req.file.filename;
     }
     exploreCat.save((error, expCat) => {
       if (error) return res.status(400).json({ error });
@@ -49,12 +49,25 @@ exports.deleteExploreCategoryById = async (req, res) => {
   try {
     const { id } = req.body;
     if (id) {
-      await ExploreCategory.deleteOne({ _id: id }).exec((error, result) => {
-        if (error) return res.status(400).json({ error });
-        if (result) {
-          res.status(202).json({ result });
-        }
-      });
+      const response = await ExploreCategory.findOne({ _id: id });
+      if (response) {
+        let newBannerImage = response?.image.replace(
+          "http://localhost:5000/public/",
+          ""
+        );
+        const imagepath1 = path.join(__dirname, "../uploads", newBannerImage);
+        fs.unlink(imagepath1, (error) => {
+          if (error) {
+            console.error(error);
+          }
+        });
+        await ExploreCategory.deleteOne({ _id: id }).exec((error, result) => {
+          if (error) return res.status(400).json({ error });
+          if (result) {
+            res.status(202).json({ result });
+          }
+        });
+      }
     } else {
       res.status(400).json({ error: "Params required" });
     }
@@ -98,7 +111,7 @@ exports.updateExploreCategory = async (req, res) => {
     };
 
     if (req.file) {
-      exploreCategory.image = req.file.location;
+      exploreCategory.image = process.env.API + "/public/" + req.file.filename;
     }
 
     if (imageTitle != undefined) {
